+++ packages/main/src/webComponents/Popover/index.tsx
@@ -1,10 +1,11 @@
 import { Event } from '@ui5/webcomponents-react-base';
 import UI5Popover from '@ui5/webcomponents/dist/Popover';
-import React, { Component, CSSProperties, ReactNode, RefObject } from 'react';
+import React, { CSSProperties, ReactNode, RefObject, useCallback, useEffect, useRef } from 'react';
 import { withWebComponent, WithWebComponentPropTypes } from '../../internal/withWebComponent';
 import { PlacementType } from '../../lib/PlacementType';
 import { PopoverHorizontalAlign } from '../../lib/PopoverHorizontalAlign';
 import { PopoverVerticalAlign } from '../../lib/PopoverVerticalAlign';
+import { Ui5PopoverDomRef } from '../../interfaces/Ui5PopoverDomRef';
 
 export interface PopoverPropTypes extends WithWebComponentPropTypes {
   initialFocus?: string; // @generated
@@ -31,73 +32,51 @@ export interface PopoverPropTypes extends WithWebComponentPropTypes {
 
 const InternalPopover = withWebComponent<PopoverPropTypes>(UI5Popover);
 
-class Popover extends Component<PopoverPropTypes> {
-  private popoverRef: any;
-  private openByRef: RefObject<HTMLDivElement> = React.createRef();
+export const Popover = React.forwardRef((props: PopoverPropTypes, popoverRef: RefObject<Ui5PopoverDomRef>) => {
+  const { openBy, openByStyle, open, ...rest } = props;
 
-  static defaultProps = {
-    initialFocus: null, // @generated
-    headerText: '', // @generated
-    placementType: PlacementType.Right, // @generated
-    horizontalAlign: PopoverHorizontalAlign.Center, // @generated
-    verticalAlign: PopoverVerticalAlign.Center // @generated
+  const localPopoverRef: RefObject<Ui5PopoverDomRef> = useRef(null);
+  const openByRef: RefObject<HTMLDivElement> = useRef(null);
+  const getPopoverRef = () => popoverRef || localPopoverRef;
+
+  const handleOpenPopover = useCallback(() => {
+    return getPopoverRef().current.openBy ? getPopoverRef().current.openBy(openByRef.current) : null;
+  }, [getPopoverRef(), openByRef]);
+
+  const closePopover = () => {
+    return getPopoverRef().current.close ? getPopoverRef().current.close() : null;
   };
 
-  private handleOpenPopover = () => {
-    return this.popoverRef.openBy ? this.popoverRef.openBy(this.openByRef.current) : null;
-  };
-
-  private closePopover = () => {
-    return this.popoverRef.close ? this.popoverRef.close() : null;
-  };
-
-  getPopoverRef = (el: any) => {
-    const { innerComponentRef } = this.props;
-    this.popoverRef = el;
-    if (innerComponentRef) {
-      innerComponentRef(el);
-    }
-  };
-
-  componentDidMount() {
-    const { open } = this.props;
-    if (!this.popoverRef) {
-      return;
-    }
+  useEffect(() => {
     if (open) {
-      this.handleOpenPopover();
+      handleOpenPopover();
     } else {
-      this.closePopover();
+      closePopover();
     }
+  }, [open]);
+
+  let style = { display: 'inline-block' };
+  if (openByStyle) {
+    style = Object.assign(openByStyle, style);
   }
 
-  componentDidUpdate(prevProps) {
-    if (prevProps.open !== this.props.open) {
-      if (this.props.open) {
-        this.handleOpenPopover();
-      } else {
-        this.closePopover();
-      }
-    }
-  }
+  return (
+    <>
+      {openBy && (
+        <div style={style} onClick={handleOpenPopover} ref={openByRef}>
+          {openBy}
+        </div>
+      )}
+      <InternalPopover {...rest} ref={getPopoverRef()} />
+    </>
+  );
+});
 
-  render() {
-    const { openBy, openByStyle, ...props } = this.props;
-    let style = { display: 'inline-block' };
-    if (openByStyle) {
-      style = Object.assign(openByStyle, style);
-    }
-    return (
-      <>
-        {openBy && (
-          <div style={style} onClick={this.handleOpenPopover} ref={this.openByRef}>
-            {openBy}
-          </div>
-        )}
-        <InternalPopover {...props} innerComponentRef={this.getPopoverRef} />
-      </>
-    );
-  }
-}
-
-export { Popover };
+Popover.defaultProps = {
+  initialFocus: null, // @generated
+  headerText: '', // @generated
+  placementType: PlacementType.Right, // @generated
+  horizontalAlign: PopoverHorizontalAlign.Center, // @generated
+  verticalAlign: PopoverVerticalAlign.Center // @generated
+};
+Popover.displayName = 'Popover';